requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.1
flake8==7.0.0
grpcio==1.60.1
grpcio-tools==1.46.5
h11==0.14.0
idna==3.6
loguru==0.7.2
mccabe==0.7.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
protobuf==3.19.5
pycodestyle==2.11.1
pydantic==2.6.0
pydantic_core==2.16.1
pyflakes==3.2.0
python-multipart==0.0.7
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sentry-sdk==1.40.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
